+++ crates/cli/tests/reference/targets-target-no-modules-atomics.js
@@ -68,7 +68,7 @@ let wasm_bindgen = (function(exports) {
         wasmModule = module;
         cachedUint8ArrayMemory0 = null;
         if (typeof thread_stack_size !== 'undefined' && (typeof thread_stack_size !== 'number' || thread_stack_size === 0 || thread_stack_size % 65536 !== 0)) {
-            throw 'invalid stack size';
+            throw new Error('invalid stack size');
         }
         wasm.__wbindgen_start(thread_stack_size);
         return wasm;