+++ crates/cli/tests/reference/targets-target-web-experimental-reset-state-function-atomics.js
@@ -83,7 +83,7 @@ function __wbg_finalize_init(instance, module, thread_stack_size) {
     wasmModule = module;
     cachedUint8ArrayMemory0 = null;
     if (typeof thread_stack_size !== 'undefined' && (typeof thread_stack_size !== 'number' || thread_stack_size === 0 || thread_stack_size % 65536 !== 0)) {
-        throw 'invalid stack size';
+        throw new Error('invalid stack size');
     }
     wasm.__wbindgen_start(thread_stack_size);
     return wasm;